+++ husky.config.js
@@ -1,1 +1,8 @@
-module.exports = require('@jupiterone/integration-sdk-dev-tools/config/husky');
+const { hooks } = require('@jupiterone/integration-sdk-dev-tools/config/husky');
+
+module.exports = {
+  hooks: {
+    'pre-commit': `${hooks['pre-commit']} && yarn precommit`,
+    'pre-push': hooks['pre-push'],
+  },
+};